refactor(hover): keyword'ы — synthetic KeywordSymbol + ReferenceFinder + Builder
Замечание @nixel2007: разгрузить HoverProvider от keyword-специфики.
Принцип работы провайдера — три строчки: резолв ссылки → поиск
маркап-билдера → делегирование. Keyword-обработка по этому же flow
через synthetic-символ (аналог AnnotationSymbol/AnnotationReferenceFinder
из PR #3906).

- KeywordSymbol (context.symbol) — synthetic Symbol с name, готовым
  локализованным description и selectionRange. SymbolKind = Operator.
  В symbol-tree не входит (accept = no-op).
- KeywordReferenceFinder (references, @Order 190): по позиции находит
  terminal-node, фильтрует не-keyword токены (IDENTIFIER/STRING/DECIMAL/
  FLOAT/DATETIME/EOF), определяет parent-context (Функция/Процедура/Перем)
  обходом AST, и через GlobalScopeProvider.findKeywordDescription строит
  KeywordSymbol + Reference. Логика isKeywordToken / findKeywordParentContext
  переехала сюда из HoverProvider без изменений.
- KeywordSymbolMarkupContentBuilder (hover): оборачивает уже выбранное
  описание в markdown ```bsl\n<keyword>\n```\n\n_<label>_\n\n<description>.
  Resource bundle KeywordSymbolMarkupContentBuilder_{ru,en}.properties
  перенесён из providers/HoverProvider_*.properties (там был только
  keywordLabel — больше провайдер ресурсов не запрашивает).
- HoverProvider: удалены tryKeywordHover, findKeywordParentContext,
  isKeywordToken, инжекции GlobalScopeProvider/LanguageServerConfiguration/
  Resources. Остался только single-flow: resolve reference → find builder
  → delegate.
