+++ src/main/resources/com/github/_1c_syntax/bsl/languageserver/hover/KeywordSymbolMarkupContentBuilder_en.properties
